+++ src/input/input.stories.ts
@@ -1,0 +1,20 @@
+import { storiesOf, moduleMetadata } from "@storybook/angular";
+import { action } from "@storybook/addon-actions";
+import { withKnobs, boolean } from "@storybook/addon-knobs/angular";
+
+import { InputModule } from "../";
+
+storiesOf("Input", module).addDecorator(
+	moduleMetadata({
+		imports: [InputModule]
+	})
+)
+	.addDecorator(withKnobs)
+	.add("Label", () => ({
+		template: `
+		<ibm-label>
+			Some Title
+			<input type="text" class="bx--text-input" placeholder="Optional placeholder text">
+		</ibm-label>
+	`
+	}));
