+++ src/input/label.component.spec.ts
@@ -2,7 +2,7 @@ import { BrowserAnimationsModule } from "@angular/platform-browser/animations";
 import { ComponentFixture, TestBed, fakeAsync, tick, async } from "@angular/core/testing";
 import { By } from "@angular/platform-browser";
 import { DebugElement } from "@angular/core";
-import { StaticIconModule } from "./../icon/static-icon.module";
+import { StaticIconModule } from "../icon/static-icon.module";
 
 import { LabelComponent } from "./label.component";
