+++ src/radio/radio.component.spec.ts
@@ -9,7 +9,7 @@ import {
 import { By } from "@angular/platform-browser";
 import { FormsModule } from "@angular/forms";
 import { DebugElement, Component } from "@angular/core";
-import { StaticIconModule } from "./../icon/static-icon.module";
+import { StaticIconModule } from "../icon/static-icon.module";
 
 import { RadioComponent, RadioGroup } from "./radio.component";
