+++ src/switch/switch.stories.ts
@@ -1,0 +1,18 @@
+import { storiesOf, moduleMetadata } from "@storybook/angular";
+import { action } from "@storybook/addon-actions";
+import { withKnobs, boolean } from "@storybook/addon-knobs/angular";
+
+import { SwitchModule } from "../";
+
+storiesOf("Switch", module).addDecorator(
+	moduleMetadata({
+		imports: [SwitchModule]
+	})
+)
+	.addDecorator(withKnobs)
+	.add("Basic", () => ({
+		template: `<ibm-switch [disabled]="disabled"></ibm-switch>`,
+		props: {
+			disabled: boolean("disabled", false),
+		}
+	}));
